# apps/auth-service/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Auth Service API",
    "description": "API documentation for the authentication service",
    "version": "1.0.0"
  },
  "host": "localhost:6001",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user-registration": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "Rahim Badsa"
                },
                "email": {
                  "example": "rahimbadsa723@gmail.com"
                },
                 "password": {
                  "example": "12345678"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/verify-user": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "rahimbadsa723@gmail.com"
                },
                "otp": {
                  "example": ""
                },
                "password": {
                  "example": "1235678"
                },
                "name": {
                  "example": "Rahim Badsa"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth-test": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}